+++ src/main/java/kitchenpos/application/OrderService.java
@@ -1,8 +1,8 @@
 package kitchenpos.application;
 
-import kitchenpos.application.request.ChangeOrderStatusRequest;
-import kitchenpos.application.request.OrderRequest;
-import kitchenpos.application.response.OrderResponse;
+import kitchenpos.application.request.order.ChangeOrderStatusRequest;
+import kitchenpos.application.request.order.OrderRequest;
+import kitchenpos.application.response.order.OrderResponse;
 import kitchenpos.application.response.ResponseAssembler;
 import kitchenpos.dao.MenuDao;
 import kitchenpos.dao.OrderDao;
