+++ src/main/java/kitchenpos/application/request/order/OrderRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.application.request;
+package kitchenpos.application.request.order;
 
 import java.util.List;
