+++ src/main/java/kitchenpos/application/response/menu/MenuProductResponse.java
@@ -1,4 +1,4 @@
-package kitchenpos.application.response;
+package kitchenpos.application.response.menu;
 
 public class MenuProductResponse {
