+++ src/test/java/kitchenpos/application/MenuServiceTest.java
@@ -1,8 +1,8 @@
 package kitchenpos.application;
 
-import kitchenpos.application.request.MenuProductRequest;
-import kitchenpos.application.request.MenuRequest;
-import kitchenpos.application.response.MenuResponse;
+import kitchenpos.application.request.menu.MenuProductRequest;
+import kitchenpos.application.request.menu.MenuRequest;
+import kitchenpos.application.response.menu.MenuResponse;
 import kitchenpos.dao.MenuDao;
 import kitchenpos.dao.MenuGroupDao;
 import kitchenpos.dao.MenuProductDao;
