+++ src/test/java/kitchenpos/application/OrderServiceTest.java
@@ -1,9 +1,9 @@
 package kitchenpos.application;
 
-import kitchenpos.application.request.ChangeOrderStatusRequest;
-import kitchenpos.application.request.OrderLineItemRequest;
-import kitchenpos.application.request.OrderRequest;
-import kitchenpos.application.response.OrderResponse;
+import kitchenpos.application.request.order.ChangeOrderStatusRequest;
+import kitchenpos.application.request.order.OrderLineItemRequest;
+import kitchenpos.application.request.order.OrderRequest;
+import kitchenpos.application.response.order.OrderResponse;
 import kitchenpos.dao.MenuDao;
 import kitchenpos.dao.OrderDao;
 import kitchenpos.dao.OrderLineItemDao;
